Move the registration form into its own component as the code is quite long by itself
