+++ frontend/src/components/RegisterForm.css
@@ -1,0 +1,7 @@
+.register__long {
+  width: 100vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 2em 0;
+}
